abstracted value to ObservableValue for reuse
